CDA: add vite-plugin-dreamland to tsx/jsx templates

+++ create-dreamland-app/src/scaffold.ts
@@ -11,25 +11,27 @@ interface options {
 
 async function scaffold(opts: options) {
     if (opts.scaffoldType === "tsx/jsx") {
-        try {
-            await downloadTemplate('github:motortruck1221/dreamland-stuff/create-dreamland-app/templates', {
-                force: true,
-                provider: 'github',
-                cwd: opts.projectName,
-                dir: '.'
-            })
-        } catch(err: any) {
-            //remove the dir if it's likely to be created by the CLI 
-            if (opts.projectName !== '.' && opts.projectName !== './' && opts.projectName.startsWith('../')) {
-                try {
-                    fs.rmdirSync(opts.projectName);
-                } catch (_) {}
-            }
-            if (err.message.includes('404')) {
-                throw new Error('It looks like we were not able to get the template. \n Please try again later');
-            }
-            else {
-                throw new Error(err.message);
+        if (opts.tsScaffold === true) {
+            try {
+                await downloadTemplate('github:motortruck1221/dreamland-stuff/create-dreamland-app/templates/tsx', {
+                    force: true,
+                    provider: 'github',
+                    cwd: opts.projectName,
+                    dir: '.'
+                })
+            } catch(err: any) {
+                //remove the dir if it's likely to be created by the CLI 
+                if (opts.projectName !== '.' && opts.projectName !== './' && opts.projectName.startsWith('../')) {
+                    try {
+                        fs.rmdirSync(opts.projectName);
+                    } catch (_) {}
+                }
+                if (err.message.includes('404')) {
+                    throw new Error('It looks like we were not able to get the template. \n Please try again later');
+                }
+                else {
+                    throw new Error(err.message);
+                }
             }
         }
